+++ packages/harness/src/index.ts
@@ -5,7 +5,8 @@ export { runHarness } from "./pipeline.js";
 
 export { InlineContextProvider } from "./context/inline.js";
 export { StaticPlanner } from "./planners/static.js";
-export { AssertionCritic } from "./critics/assertion.js";
+export { AssertionCritic, checkAssertion } from "./critics/assertion.js";
+export { LlmCritic, summarizeEvidence } from "./critics/llm.js";
 export { ConsoleReporter } from "./reporters/console.js";
 export { JsonReporter } from "./reporters/json.js";
 export { FakeDriver } from "./drivers/fake.js";